backend/requirements.txt
# Core dependencies
openai==1.78.1
gunicorn==21.2.0
setuptools==68.2.2
pillow==10.2.0
python-dotenv==1.0.0
requests==2.31.0

# Flask and extensions
flask==2.3.3
flask-cors==4.0.0
flask-session==0.5.0

# Data processing
pandas==2.1.1
scikit-learn==1.3.2
numpy>=1.24.0
openpyxl==3.1.2


# Caching and session management
python-dateutil==2.8.2
redis==5.0.1
cachetools==5.3.2
